[linkDetection] - moved sanitizeXML function to utils file
