+++ server/src/input/pdfminer/pdfminer.ts
@@ -29,7 +29,6 @@ import { Color } from '../../types/DocumentRepresentation/Color';
 import { PdfminerFigure } from '../../types/PdfminerFigure';
 import { PdfminerPage } from '../../types/PdfminerPage';
 import { PdfminerText } from '../../types/PdfminerText';
-import * as utils from '../../utils';
 import * as CommandExecuter from '../../utils/CommandExecuter';
 import logger from '../../utils/Logger';
 
@@ -61,24 +60,6 @@ export function extractPages(
       .catch(({ error }) => {
         rejectXml(`PdfMiner pdf2txt.py error: ${error}`);
       });
-  });
-}
-export function sanitizeXML(xmlPath: string): Promise<string> {
-  const startTime: number = Date.now();
-  return new Promise<any>((resolve, _reject) => {
-    try {
-      // repplace with empty char everything forbidden by XML 1.0 specifications,
-      // plus the unicode replacement character U+FFFD
-      const regex = /((?:[\0-\x08\x0B\f\x0E-\x1F\uFFFD\uFFFE\uFFFF]|[\uD800-\uDBFF](?![\uDC00-\uDFFF])|(?:[^\uD800-\uDBFF]|^)[\uDC00-\uDFFF]))/g;
-      const xml: string = fs.readFileSync(xmlPath, 'utf-8');
-      const outputFilePath = utils.getTemporaryFile('.xml');
-      fs.writeFileSync(outputFilePath, xml.replace(new RegExp(regex), ' '));
-      logger.info(`Sanitize XML: ${(Date.now() - startTime) / 1000}s`);
-      resolve(outputFilePath);
-    } catch (error) {
-      logger.warn(`Error sanitizing XML ${error}`);
-      resolve(xmlPath);
-    }
   });
 }
